pyproject.toml
[project]
name = "commitvoice"
version = "0.2.0"
description = "CommitVoice: git commit summaries as tweets, LinkedIn blurbs, or timesheet lines (optional Gemini polish)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "iamrealafzal", email = "afzalmlakdawala@gmail.com" }]
dependencies = [
    "google-generativeai>=0.7.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
commitvoice = "commitvoice.cli:main"
social-updates = "commitvoice.cli:main"

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


